Refactor error dialog imports: update paths for NetworkErrorDialog in multiple components for improved structure and consistency
