+++ client/src/components/errors/NetworkErrorDialog.tsx
