+++ client/src/pages/auth/login/Login.tsx
@@ -1,20 +1,30 @@
+/* eslint-disable  @typescript-eslint/no-explicit-any */
 import { useAuth } from "@/features/contexts/AuthContext";
 import { useNavigate } from "react-router-dom";
 import LoginForm from "./LoginForm";
 import { FormType } from "../components/CustomFormField";
 import { Layout, LayoutLabel } from "../Layout";
 import { login } from "../../../features/api/services";
+import { useState } from "react";
+import NetworkErrorDialog from "@/components/errors/NetworkErrorDialog";
 
 export default function Login() {
     const { setToken } = useAuth();
     const navigate = useNavigate();
+    const [openNetworkErrorDialog, setOpenNetworkErrorDialog] = useState(false);
 
     const handleLogin = async (formData: FormType) => {
-        await login(formData)
-            .then((data) => {
-                setToken(data.token)
-                navigate('/home')
-            })
+        try {
+            const response_data = await login(formData);
+            setToken(response_data.token);
+            navigate('/home');
+        } catch (error: any) {
+            if (error.message === "Network Error") {
+                setOpenNetworkErrorDialog(true);
+            } else {
+                console.log(error);
+            }
+        }
         
     };
 
@@ -22,6 +32,7 @@ export default function Login() {
         <Layout>
             <LayoutLabel>Connexion</LayoutLabel>
             <LoginForm onLogin={handleLogin}/>
+            {openNetworkErrorDialog && <NetworkErrorDialog setIsOpen={setOpenNetworkErrorDialog} />}
         </Layout>
     )
 }
